+++ apps/e2e/tests/stack-server/api/internal-project.test.ts
@@ -1,6 +1,7 @@
 import { describe, expect, test } from "vitest";
 import request from "supertest";
-import { BASE_URL, INTERNAL_PROJECT_CLIENT_KEY, INTERNAL_PROJECT_ID } from "../helpers";
+import { BASE_URL, INTERNAL_PROJECT_CLIENT_KEY, INTERNAL_PROJECT_ID } from "../../helpers";
+import crypto from "crypto";
 
 const AUTH_HEADER = {
   "x-stack-project-id": INTERNAL_PROJECT_ID,
@@ -12,11 +13,11 @@ const JSON_HEADER = {
 }
 
 function randomString() {
-  return Math.random().toString(36);
+  return crypto.randomBytes(16).toString("hex");
 }
 
 async function signUpWithEmailPassword() {
-  const email = randomString() + "@example.com";
+  const email = randomString() + "@stack-test.example.com";
   const password = randomString();
   const response = await request(BASE_URL).post("/api/v1/auth/signup").set(AUTH_HEADER).set(JSON_HEADER).send({
     email,
@@ -42,25 +43,26 @@ describe("Various internal project tests", () => {
     expect(response.status).toBe(307);
   });
 
-  test("Test API", async () => {
+  test("API root (no authentication)", async () => {
     const response = await request(BASE_URL).get("/api/v1");
     expect(response.status).toBe(200);
     expect(response.text).contains("Stack API")
+    expect(response.text).contains("Authentication: None")
   });
 
-  test("Credential Sign Up", async () => {
+  test("Credential sign up", async () => {
     const { response } = await signUpWithEmailPassword();
     expect(response.status).toBe(200);
   });
 
-  test("Credential Sign In", async () => {
+  test("Credential sign in", async () => {
     const { email, password } = await signUpWithEmailPassword();
     const { response } = await signInWithEmailPassword(email, password);
 
     expect(response.status).toBe(200);
   });
 
-  test("Get Current User", async () => {
+  test("Get current user", async () => {
     const { email, password, response } = await signUpWithEmailPassword();
     await signInWithEmailPassword(email, password);
 
@@ -70,5 +72,21 @@ describe("Various internal project tests", () => {
     });
     expect(response2.status).toBe(200);
     expect(response2.body.primaryEmail).toBe(email)
+  });
+
+  test("API root (signed in)", async () => {
+    const { email, password, response } = await signUpWithEmailPassword();
+    await signInWithEmailPassword(email, password);
+
+    const response2 = await request(BASE_URL).get("/api/v1").set({
+      ...AUTH_HEADER,
+      'x-stack-request-type': 'client',
+      'authorization': 'StackSession ' + response.body.accessToken,
+    });
+    expect(response2.status).toBe(200);
+    expect(response2.text).contains("Stack API");
+    expect(response2.text).contains("Authentication: Client");
+    expect(response2.text).contains("Project: " + INTERNAL_PROJECT_ID);
+    expect(response2.text).contains("User: " + email);
   });
 });