+++ packages/stack-server/src/app/(main)/route.tsx
@@ -1,4 +1,4 @@
-import { deprecatedSmartRouteHandler } from "@/lib/route-handlers";
+import { deprecatedSmartRouteHandler } from "@/route-handlers/smart-route-handler";
 import { redirect } from "next/navigation";
 
 export const GET = deprecatedSmartRouteHandler(async () => {