+++ packages/stack-server/src/app/api/v1/auth/token/route.tsx
@@ -2,7 +2,8 @@ import * as yup from "yup";
 import { InvalidGrantError, Request as OAuthRequest, Response as OAuthResponse, InvalidClientError } from "@node-oauth/oauth2-server";
 import { NextRequest, NextResponse } from "next/server";
 import { oauthServer } from "@/oauth";
-import { deprecatedParseRequest, deprecatedSmartRouteHandler } from "@/lib/route-handlers";
+import { deprecatedSmartRouteHandler } from "@/route-handlers/smart-route-handler";
+import { deprecatedParseRequest } from "@/route-handlers/smart-request";
 import { KnownErrors } from "@stackframe/stack-shared";
 
 // make this specific to each grant type later